requirements.txt
-e git+ssh://git@github.com/ethan-jiang-1/fly_face.git@a6a2cfe661988397d2325a858e754814a881075b#egg=tflite_runtime_hsi&subdirectory=hsi_tflite_interpeter/tflite_custom.macos
Pillow==9.0.1
PySimpleGUI==4.57.0
absl-py==1.0.0
argcomplete==1.12.3
attrs==21.4.0
certifi==2021.10.8
click==8.0.3
cycler==0.11.0
dlib==19.23.1
fake-bpy-module-3.0==20211212
flake8==4.0.1
fonttools==4.30.0
kiwisolver==1.3.2
matplotlib==3.5.1
mccabe==0.6.1
mediapipe==0.8.9.1
numpy==1.22.3
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
packaging==21.0
pandas==1.4.1
pipx==0.16.4
protobuf==3.19.4
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
tqdm==4.63.0
userpath==1.7.0
